clean codebase except websocket
