+++ sotopia/ui/streamlit_ui/__init__.py
@@ -1,7 +1,3 @@
 from .utils import reset_database
-from .rendering.rendering_utils import (
-    _agent_profile_to_friendabove_self,
-    render_for_humans,
-)
 
-__all__ = ["reset_database", "_agent_profile_to_friendabove_self", "render_for_humans"]
+__all__ = ["reset_database"]
